Tested and resolved defects
