+++ Scripts/Misc/MondainsLegacy.cs
@@ -6,6 +6,8 @@ using Server.Engines.InstancedPeerless;
 using Server.Engines.Quests;
 using Server.Items;
 using Server.Mobiles;
+using Server.Gumps;
+using Server.Network;
 
 namespace Server
 {
@@ -209,8 +211,9 @@ namespace Server
         }
         public static void Initialize()
         {
-            CommandSystem.Register("DecorateML", AccessLevel.Administrator, new CommandEventHandler(DecorateML_OnCommand));
-            CommandSystem.Register("SettingsML", AccessLevel.Administrator, new CommandEventHandler(SettingsML_OnCommand));
+			CommandSystem.Register("DecorateML", AccessLevel.Administrator, new CommandEventHandler(DecorateML_OnCommand));
+			CommandSystem.Register("DecorateMLDelete", AccessLevel.Administrator, new CommandEventHandler(DecorateMLDelete_OnCommand));
+			CommandSystem.Register("SettingsML", AccessLevel.Administrator, new CommandEventHandler(SettingsML_OnCommand));
             CommandSystem.Register("Quests", AccessLevel.GameMaster, new CommandEventHandler(Quests_OnCommand));
 
             LoadSettings();
@@ -403,18 +406,25 @@ namespace Server
                    region.IsPartOf("Labyrinth");
         }
 
-        [Usage("DecorateML")]
+		[Usage("DecorateMLDelete")]
+		[Description("Deletes Mondain's Legacy world decoration.")]
+		private static void DecorateMLDelete_OnCommand(CommandEventArgs e)
+		{
+			WeakEntityCollection.DeleteEntities("ml");
+		}
+
+		[Usage("DecorateML")]
         [Description("Generates Mondain's Legacy world decoration.")]
         private static void DecorateML_OnCommand(CommandEventArgs e)
         {
             e.Mobile.SendMessage("Generating Mondain's Legacy world decoration, please wait.");
 
-            Decorate.Generate("Data/Mondain's Legacy/Trammel", Map.Trammel);
-            Decorate.Generate("Data/Mondain's Legacy/Felucca", Map.Felucca);
-            Decorate.Generate("Data/Mondain's Legacy/Ilshenar", Map.Ilshenar);
-            Decorate.Generate("Data/Mondain's Legacy/Malas", Map.Malas);
-            Decorate.Generate("Data/Mondain's Legacy/Tokuno", Map.Tokuno);
-            Decorate.Generate("Data/Mondain's Legacy/TerMur", Map.TerMur);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/Trammel", Map.Trammel);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/Felucca", Map.Felucca);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/Ilshenar", Map.Ilshenar);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/Malas", Map.Malas);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/Tokuno", Map.Tokuno);
+            Decorate.Generate("ml", "Data/Mondain's Legacy/TerMur", Map.TerMur);
 
             PeerlessAltar altar;
             PeerlessTeleporter tele;
@@ -425,10 +435,11 @@ namespace Server
 
             if (!FindItem(86, 1627, 0, Map.Malas, altar))
             {
-                altar = new BedlamAltar();
-                altar.MoveToWorld(new Point3D(86, 1627, 0), Map.Malas);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(86, 1627, 0), Map.Malas);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(99, 1617, 50), Map.Malas);
             }
 
@@ -437,9 +448,11 @@ namespace Server
 
             if (!FindItem(6502, 875, 0, Map.Trammel, altar))
             {
-                altar.MoveToWorld(new Point3D(6502, 875, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6502, 875, 0), Map.Trammel);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(6511, 949, 26), Map.Trammel);
             }
 
@@ -448,9 +461,11 @@ namespace Server
 
             if (!FindItem(6502, 875, 0, Map.Felucca, altar))
             {
-                altar.MoveToWorld(new Point3D(6502, 875, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6502, 875, 0), Map.Felucca);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(6511, 949, 26), Map.Felucca);
             }
 
@@ -459,9 +474,11 @@ namespace Server
 
             if (!FindItem(6511, 506, -34, Map.Trammel, altar))
             {
-                altar.MoveToWorld(new Point3D(6511, 506, -34), Map.Trammel);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6511, 506, -34), Map.Trammel);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(6518, 365, 46), Map.Trammel);
             }
 
@@ -470,9 +487,11 @@ namespace Server
 
             if (!FindItem(6511, 506, -34, Map.Felucca, altar))
             {
-                altar.MoveToWorld(new Point3D(6511, 506, -34), Map.Felucca);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6511, 506, -34), Map.Felucca);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(6518, 365, 46), Map.Felucca);
             }
 
@@ -481,30 +500,38 @@ namespace Server
 
             if (!FindItem(6509, 167, 6, Map.Trammel, altar))
             {
-                altar.MoveToWorld(new Point3D(6509, 167, 6), Map.Trammel);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6509, 167, 6), Map.Trammel);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.Visible = true;
                 tele.ItemID = 0xDDA;
                 tele.MoveToWorld(new Point3D(6501, 137, -20), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6506, 167, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6506, 167, 0), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6509, 164, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6509, 164, 0), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6506, 164, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6506, 164, 0), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6512, 167, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6512, 167, 0), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6509, 170, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6509, 170, 0), Map.Trammel);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6512, 170, 0), Map.Trammel);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6512, 170, 0), Map.Trammel);
             }
 
             // Prism of Light - Felucca
@@ -512,30 +539,38 @@ namespace Server
 
             if (!FindItem(6509, 167, 6, Map.Felucca, altar))
             {
-                altar.MoveToWorld(new Point3D(6509, 167, 6), Map.Felucca);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(6509, 167, 6), Map.Felucca);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.Visible = true;
                 tele.ItemID = 0xDDA;
                 tele.MoveToWorld(new Point3D(6501, 137, -20), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6506, 167, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6506, 167, 0), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6509, 164, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6509, 164, 0), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x581);
-                pillar.MoveToWorld(new Point3D(6506, 164, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6506, 164, 0), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6512, 167, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6512, 167, 0), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6509, 170, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6509, 170, 0), Map.Felucca);
 
                 pillar = new PrismOfLightPillar((PrismOfLightAltar)altar, 0x481);
-                pillar.MoveToWorld(new Point3D(6512, 170, 0), Map.Felucca);
+				WeakEntityCollection.Add("ml", pillar);
+				pillar.MoveToWorld(new Point3D(6512, 170, 0), Map.Felucca);
             }
 
             // The Citadel - Malas
@@ -543,9 +578,11 @@ namespace Server
 
             if (!FindItem(89, 1885, 0, Map.Malas, altar))
             {
-                altar.MoveToWorld(new Point3D(89, 1885, 0), Map.Malas);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(89, 1885, 0), Map.Malas);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(111, 1955, 0), Map.Malas);
             }
 
@@ -554,9 +591,11 @@ namespace Server
 
             if (!FindItem(2170, 1255, -60, Map.Ilshenar, altar))
             {
-                altar.MoveToWorld(new Point3D(2170, 1255, -60), Map.Ilshenar);
+				WeakEntityCollection.Add("ml", altar);
+				altar.MoveToWorld(new Point3D(2170, 1255, -60), Map.Ilshenar);
                 tele = new PeerlessTeleporter(altar);
-                tele.PointDest = altar.ExitDest;
+				WeakEntityCollection.Add("ml", tele);
+				tele.PointDest = altar.ExitDest;
                 tele.MoveToWorld(new Point3D(2139, 1271, -57), Map.Ilshenar);
             }
 
@@ -565,7 +604,8 @@ namespace Server
 
             if (!FindItem(363, 157, 5, Map.TerMur, sAltar))
             {
-                sAltar.MoveToWorld(new Point3D(363, 157, 0), Map.TerMur);
+				WeakEntityCollection.Add("ml", sAltar);
+				sAltar.MoveToWorld(new Point3D(363, 157, 0), Map.TerMur);
 
             }
 
@@ -574,7 +614,8 @@ namespace Server
 
             if (!FindItem(822, 756, 56, Map.TerMur, mAltar))
             {
-                mAltar.MoveToWorld(new Point3D(822, 756, 56), Map.TerMur);
+				WeakEntityCollection.Add("ml", sAltar);
+				mAltar.MoveToWorld(new Point3D(822, 756, 56), Map.TerMur);
             }
 
             e.Mobile.SendMessage("Mondain's Legacy world generating complete.");
@@ -604,4 +645,112 @@ namespace Server
                 }));
         }
     }
+
+	public class MondainsLegacyGump : Gump
+	{
+		public MondainsLegacyGump()
+			: base(50, 50)
+		{
+			this.Closable = true;
+			this.Disposable = true;
+			this.Dragable = true;
+			this.Resizable = false;
+
+			this.AddPage(0);
+			this.AddBackground(0, 0, 308, 390, 0x2454);
+
+			// title
+			this.AddLabel(125, 10, 150, "Settings");
+			this.AddImage(256, 5, 0x9E1);
+
+			// dungeons			
+			this.AddButton(20, 60, MondainsLegacy.PalaceOfParoxysmus ? 0x939 : 0x938, MondainsLegacy.PalaceOfParoxysmus ? 0x939 : 0x938, 1, GumpButtonType.Reply, 0);
+			this.AddButton(20, 85, MondainsLegacy.TwistedWeald ? 0x939 : 0x938, MondainsLegacy.TwistedWeald ? 0x939 : 0x938, 2, GumpButtonType.Reply, 0);
+			this.AddButton(20, 110, MondainsLegacy.BlightedGrove ? 0x939 : 0x938, MondainsLegacy.BlightedGrove ? 0x939 : 0x938, 3, GumpButtonType.Reply, 0);
+			this.AddButton(20, 135, MondainsLegacy.Bedlam ? 0x939 : 0x938, MondainsLegacy.Bedlam ? 0x939 : 0x938, 4, GumpButtonType.Reply, 0);
+			this.AddButton(20, 160, MondainsLegacy.PrismOfLight ? 0x939 : 0x938, MondainsLegacy.PrismOfLight ? 0x939 : 0x938, 5, GumpButtonType.Reply, 0);
+			this.AddButton(20, 185, MondainsLegacy.Citadel ? 0x939 : 0x938, MondainsLegacy.Citadel ? 0x939 : 0x938, 6, GumpButtonType.Reply, 0);
+			this.AddButton(20, 210, MondainsLegacy.PaintedCaves ? 0x939 : 0x938, MondainsLegacy.PaintedCaves ? 0x939 : 0x938, 7, GumpButtonType.Reply, 0);
+			this.AddButton(20, 235, MondainsLegacy.Labyrinth ? 0x939 : 0x938, MondainsLegacy.Labyrinth ? 0x939 : 0x938, 8, GumpButtonType.Reply, 0);
+			this.AddButton(20, 260, MondainsLegacy.Sanctuary ? 0x939 : 0x938, MondainsLegacy.Sanctuary ? 0x939 : 0x938, 9, GumpButtonType.Reply, 0);
+			this.AddButton(20, 285, MondainsLegacy.StygianDragonLair ? 0x939 : 0x938, MondainsLegacy.StygianDragonLair ? 0x939 : 0x938, 10, GumpButtonType.Reply, 0);
+			this.AddButton(20, 310, MondainsLegacy.MedusasLair ? 0x939 : 0x938, MondainsLegacy.MedusasLair ? 0x939 : 0x938, 11, GumpButtonType.Reply, 0);
+			this.AddButton(20, 335, MondainsLegacy.Spellweaving ? 0x939 : 0x938, MondainsLegacy.Spellweaving ? 0x939 : 0x938, 12, GumpButtonType.Reply, 0);
+			this.AddButton(20, 360, MondainsLegacy.PublicDonations ? 0x939 : 0x938, MondainsLegacy.PublicDonations ? 0x939 : 0x938, 13, GumpButtonType.Reply, 0);
+
+			this.AddLabel(45, 56, 0x226, "Palace of Paroxysmus");
+			this.AddLabel(45, 81, 0x226, "Twisted Weald");
+			this.AddLabel(45, 106, 0x226, "Blighted Grove");
+			this.AddLabel(45, 131, 0x226, "Bedlam");
+			this.AddLabel(45, 156, 0x226, "Prism of Light");
+			this.AddLabel(45, 181, 0x226, "The Citadel");
+			this.AddLabel(45, 206, 0x226, "Painted Caves");
+			this.AddLabel(45, 231, 0x226, "Labyrinth");
+			this.AddLabel(45, 256, 0x226, "Sanctuary");
+			this.AddLabel(45, 281, 0x226, "StygianDragonLair");
+			this.AddLabel(45, 306, 0x226, "MedusasLair");
+			this.AddLabel(45, 331, 0x226, "Spellweaving");
+			this.AddLabel(45, 356, 0x226, "PublicDonations");
+
+			// legend
+			this.AddLabel(243, 205, 0x226, "Legend:");
+
+			this.AddImage(218, 235, 0x938);
+			this.AddLabel(243, 231, 0x226, "disabled");
+			this.AddImage(218, 260, 0x939);
+			this.AddLabel(243, 256, 0x226, "enabled");
+		}
+
+		public override void OnResponse(NetState sender, RelayInfo info)
+		{
+			switch (info.ButtonID)
+			{
+				case 0:
+					MondainsLegacy.SaveSetings();
+					break;
+				case 1:
+					MondainsLegacy.PalaceOfParoxysmus = !MondainsLegacy.PalaceOfParoxysmus;
+					break;
+				case 2:
+					MondainsLegacy.TwistedWeald = !MondainsLegacy.TwistedWeald;
+					break;
+				case 3:
+					MondainsLegacy.BlightedGrove = !MondainsLegacy.BlightedGrove;
+					break;
+				case 4:
+					MondainsLegacy.Bedlam = !MondainsLegacy.Bedlam;
+					break;
+				case 5:
+					MondainsLegacy.PrismOfLight = !MondainsLegacy.PrismOfLight;
+					break;
+				case 6:
+					MondainsLegacy.Citadel = !MondainsLegacy.Citadel;
+					break;
+				case 7:
+					MondainsLegacy.PaintedCaves = !MondainsLegacy.PaintedCaves;
+					break;
+				case 8:
+					MondainsLegacy.Labyrinth = !MondainsLegacy.Labyrinth;
+					break;
+				case 9:
+					MondainsLegacy.Sanctuary = !MondainsLegacy.Sanctuary;
+					break;
+				case 10:
+					MondainsLegacy.StygianDragonLair = !MondainsLegacy.StygianDragonLair;
+					break;
+				case 11:
+					MondainsLegacy.MedusasLair = !MondainsLegacy.MedusasLair;
+					break;
+				case 12:
+					MondainsLegacy.Spellweaving = !MondainsLegacy.Spellweaving;
+					break;
+				case 13:
+					MondainsLegacy.PublicDonations = !MondainsLegacy.PublicDonations;
+					break;
+			}
+
+			if (info.ButtonID > 0)
+				sender.Mobile.SendGump(new MondainsLegacyGump());
+		}
+	}
 }
